Add a safe version that doesn't use libfringe; 'fringe' feature enables old behaviour
