+++ README.md
@@ -10,14 +10,11 @@ Turn serde+bincode into a pipe: push `T`s and pull `u8`s, or vice versa.
 
 This library gives you a `Serializer` pipe, into which you can push `T`s and pull `u8`s; and a `Deserializer` pipe, into which you can push `u8`s and pull `T`s.
 
-Both are bounded in their memory usage, i.e. they do not simply allocate a vector for serde+bincode to serialize into. Instead, [libfringe](https://github.com/edef1c/libfringe) is leveraged to turn serde+bincode into a Generator from which `u8`s can be pulled from/pushed to on demand.
-
-This is useful for example if you have 10GiB memory available, and want to serialize+send or receive+deserialize an 8GiB vector. Note, this is perfectly possible with serde+bincode normally, but you'd need to dedicate a thread that blocks until completion to the task.
+This by default works by allocating a vector to hold the intermediate `u8`s. However the `fringe` feature can be enabled, which uses [libfringe](https://github.com/edef1c/libfringe) to turn serde+bincode into a Generator, resulting in bounded memory usage.
 
 ## Example
 
 ```rust
-extern crate serde_pipe;
 use serde_pipe::Serializer;
 
 let large_vector = (0..1u64<<30).collect::<Vec<_>>();
@@ -32,7 +29,7 @@ while let Some(pull) = serializer.pull() {
 
 ## Note
 
-This crate currently depends on [libfringe](https://github.com/edef1c/libfringe), and as such inherits these limitations:
+The `fringe` feature depends on [libfringe](https://github.com/edef1c/libfringe), and so enabling it inherits these limitations:
  * Rust nightly is required for the `asm` and `naked_functions` features;
  * The architectures currently supported are: x86, x86_64, aarch64, or1k;
  * The platforms currently supported are: bare metal, Linux (any libc), FreeBSD, DragonFly BSD, macOS. Windows is not supported.
